SCRAPE DATA,DISPLAY DATA
A frontend was set up to grab the data that was sent from the front end. This includes creating a layout o fhow the frontend is designed and forming the data to be sent to the server. The client folder also connects with the database allowing the data to be saved whenever the information is submitted. On the server side, a function was created to look through the html file to find all instaces of a tags. These url will be put into an array to be used for later.